pyproject.toml
[project]
name = "ai-safety-lab"
version = "0.1.0"
description = "Courtroom-style AI safety evaluation lab for the UNICC capstone."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.49.0",
  "google-generativeai>=0.8.4",
  "openai>=1.68.2",
  "pydantic>=2.11.1",
  "pydantic-settings>=2.8.1",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "reportlab>=4.3.1",
  "streamlit>=1.43.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
